# apps/api/internal/modules/ml/handler.go
package ml

import (
	"bytes"
	"encoding/json"
	"math"
	"net/http"
	"time"

	"github.com/HesandaLiyanage/Waypoint-Group/apps/api/internal/platform/response"
)

type Handler struct {
	mlServiceURL string
	httpClient   *http.Client
}

func NewHandler(mlServiceURL string) *Handler {
	return &Handler{
		mlServiceURL: mlServiceURL,
		httpClient:   &http.Client{Timeout: 5 * time.Second},
	}
}

type EtaRequest struct {
	OriginLat      float64 `json:"origin_lat"`
	OriginLng      float64 `json:"origin_lng"`
	DestinationLat float64 `json:"destination_lat"`
	DestinationLng float64 `json:"destination_lng"`
}

type EtaResponse struct {
	DurationMinutes  float64   `json:"duration_minutes"`
	DistanceKM       float64   `json:"distance_km"`
	ConfidenceScore  float64   `json:"confidence_score"`
	EstimatedArrival time.Time `json:"estimated_arrival"`
}

func (h *Handler) PredictEta(w http.ResponseWriter, r *http.Request) {
	var req EtaRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "Invalid JSON body")
		return
	}

	// Try proxying to Python FastAPI ML service first
	if h.mlServiceURL != "" {
		body, _ := json.Marshal(req)
		resp, err := h.httpClient.Post(h.mlServiceURL+"/api/v1/predict/eta", "application/json", bytes.NewBuffer(body))
		if err == nil && resp.StatusCode == http.StatusOK {
			defer resp.Body.Close()
			var result EtaResponse
			if err := json.NewDecoder(resp.Body).Decode(&result); err == nil {
				response.JSON(w, http.StatusOK, result)
				return
			}
		}
	}

	// Fallback calculation using Haversine formula and average speed 40 km/h (typical Sri Lanka road average)
	dist := haversine(req.OriginLat, req.OriginLng, req.DestinationLat, req.DestinationLng)
	durationHours := dist / 40.0
	durationMins := math.Round(durationHours * 60)
	if durationMins < 5 {
		durationMins = 5
	}

	eta := time.Now().Add(time.Duration(durationMins) * time.Minute)

	response.JSON(w, http.StatusOK, EtaResponse{
		DurationMinutes:  durationMins,
		DistanceKM:       math.Round(dist*100) / 100,
		ConfidenceScore:  0.88,
		EstimatedArrival: eta,
	})
}

func haversine(lat1, lon1, lat2, lon2 float64) float64 {
	const earthRadius = 6371.0 // km
	dLat := (lat2 - lat1) * (math.Pi / 180.0)
	dLon := (lon2 - lon1) * (math.Pi / 180.0)

	rLat1 := lat1 * (math.Pi / 180.0)
	rLat2 := lat2 * (math.Pi / 180.0)

	a := math.Sin(dLat/2)*math.Sin(dLat/2) +
		math.Cos(rLat1)*math.Cos(rLat2)*math.Sin(dLon/2)*math.Sin(dLon/2)
	c := 2 * math.Atan2(math.Sqrt(a), math.Sqrt(1-a))

	return earthRadius * c
}